+++ pkg/reconciler/pendingpipelinerun/util.go
@@ -1,4 +1,4 @@
-package tektonwrapper
+package pendingpipelinerun
 
 import (
 	"context"
